# sirius-api/contracts/openapi.v1.baseline.yaml
openapi: 3.0.3
info:
  title: Sirius Community API
  description: |
    Versioned Community contract for the live Fiber `/api/v1` surface.

    Auth: every `/api/v1` operation requires the `X-API-Key` header. The only
    unauthenticated route in Community is `GET /health` (outside this document).

    Correlation: Fiber `requestid` middleware (default config) uses the
    `X-Request-ID` header. When the client supplies the header it is echoed on
    the response; otherwise the server generates an ID with Fiber
    `utils.UUID` — an RFC 4122 version-4 UUID-formatted string from a seeded
    monotonic counter (not cryptographically random `google/uuid`). If seed
    initialization fails, the nil UUID
    `00000000-0000-0000-0000-000000000000` may be returned. The ID is stored in
    Fiber locals under the key `requestid` and must appear on success and error
    responses alike.

    Error shapes (current, inconsistent across handlers):
      - `{ "error": "<string>" }` — most handlers and API-key middleware
      - `{ "error": "<string>", "details": "<string>" }` — some event handlers
      - `{ "success": false, "error": "<string>" }` — scan control handlers

    Canonical future error shape (not yet universally implemented):
      `{ "error": { "code": "<string>", "message": "<string>", "details": {} },
         "request_id": "<string>" }`

    Reserved namespaces (policy only; Community does not register handlers):
      - `/api/pro/v1` — Pro edition routes
...

...
servers:
...
paths:
...
  /api/v1/logs:
    get:
      tags: [logs]
      operationId: getLogs
      summary: Retrieve stored logs
      description: Internal service log surface.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: Log list payload
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        'default':
          $ref: '#/components/responses/Error'
    post:
      tags: [logs]
      operationId: submitLog
      summary: Submit a log entry
      description: Internal service log surface.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogSubmissionRequest'
            examples:
              valid:
                value:
                  service: sirius-ui
                  level: info
                  message: user opened dashboard
      responses:
        '201':
          description: Log accepted
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                type: object
                required: [message, log_id]
                properties:
                  message:
                    type: string
                  log_id:
                    type: string
        '400':
          $ref: '#/components/responses/ErrorWithDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ErrorWithDetails'
...
  /api/v1/statistics/vulnerability-trends:
    get:
      tags: [statistics]
      operationId: getVulnerabilityTrends
      summary: Vulnerability trend series
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: Trend payload
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Infrastructure root key (SIRIUS_API_KEY_FILE / SIRIUS_API_KEY) or a
        Valkey-backed user-generated API key.
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
      description: |
        Optional client correlation ID. Echoed when present; otherwise generated by
        Fiber `utils.UUID` (RFC4122 v4-formatted seeded counter).
  headers:
    RequestID:
      description: |
        Request correlation ID (echoed client value or Fiber `utils.UUID` generated
        value). Present on success and error responses.
      schema:
        type: string
  schemas:
    LogSubmissionRequest:
      type: object
      required: [service, level, message]
      properties:
        service:
          type: string
        subcomponent:
          type: string
        level:
          type: string
          enum: [debug, info, warn, error]
        message:
          type: string
        metadata:
          type: object
          additionalProperties: true
        context:
          type: object
          additionalProperties: true
    ErrorMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
      description: Dominant current error shape.
    ErrorWithDetails:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
      description: Current event-handler error shape.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Error:
      description: Handler error using the dominant `{error}` shape
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    ErrorWithDetails:
      description: Handler error that may include details
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorWithDetails'
...